requirements.txt
openspg-kag==0.6

filelock

fastapi
sse_starlette
uvicorn

